Create README.md
I am a creative Logo maker, creative custom QR generator and an English Translator bilingual with 2 years of experience with a strong background in translating various documents in English, Hindi, Bangla, and Gujrati. My first priority will be to satisfy clients and deliver work assignments on time to build strong relationships for the future. Being Indian and with an excellent command of these languages and a track record of impeccable proofreading, editing, and translating. Adept in accurately assessing the context of the material, and translating it understandably and appropriately. Bringing forth a positive work ethic and a commitment to efficiency and integrity.